Move compound to under ethereum.
